Added Lua script to generate a custom theme.
With a command  hueShift(1, 1, 1, 1)  , which is located at the end of a script, you now can be able to hue shift the original color scheme of a theme. So you can be able to make your custom-colored theme. First three numbers are R, G, B, and the last one is a general color: black, white, 0-9. 1 is original black, 8 is original white. 0-9 would be absolute black. Executing the script creates a theme file in the location of your Lua executor. You can google and download "Lua for Windows" to perform this.